Added stable* tracking GUI with calibration module
This is a huge change since the last version. It is basically a full featured program now that allows the user to perform a 3D calibration with two views of the object-tracking space, perform the tracking and reconstruct the trajectory in 3D. At the moment however it is only able to track one object at a time.